+++ components/links/message/index.tsx
@@ -5,9 +5,9 @@ interface IMessageLinkProps {
     messageId: number,
     messageLink: string,
     messageImagePath: string,
+    messageImageSize: number,
     messageAuthorType: number,
     messageAuthorIsStatus: boolean,
-    messageAuthorIsOnline: boolean,
     messageAuthorName: string,
     messageContent: string,
     messageTime: string,
@@ -25,7 +25,7 @@ const MessageLink = (props: IMessageLinkProps) => {
                     <Avatar
                         name={props.messageAuthorName}
                         imagePath={props.messageImagePath}
-                        size={60}
+                        size={props.messageImageSize}
                         isStatus={props.messageAuthorIsStatus}
                     />
                     <div className="basis-2/4 ml-4 font-bold truncate">
@@ -33,11 +33,6 @@ const MessageLink = (props: IMessageLinkProps) => {
                         <p className="text-sm text-gray-500">{props.messageContent}</p>
                     </div>
                     <div className="basis-1/4 flex justify-end items-center">
-                        {props.messageAuthorIsOnline &&
-                            <span className="relative flex h-3 w-3" title='Online'>
-                                <span className="animate-ping absolute inline-flex h-full w-full rounded-full bg-green-400 opacity-75"></span>
-                                <span className="relative inline-flex rounded-full h-3 w-3 bg-green-500"></span>
-                            </span>}
                         <p className="ml-2 text-md text-white font-bold">{props.messageTime}</p>
                     </div>
                 </div>
